% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFromRegion_function.R
\name{buildFromRegion}
\alias{buildFromRegion}
\title{Build empty list using genomic region.}
\usage{
buildFromRegion(chr, start, stop)
}
\arguments{
\item{chr}{Chromosome}

\item{start}{Start position}

\item{stop}{Stop position}
}
\description{
\code{buildFromRegion} returns an empty 'mgl' list given a genomic region.
}
\details{
This is one of three functions that can be used to set up the list structure.  It starts
  with a position range.
}
\examples{
\dontrun{buildFromRegion(chr = 2, start = 102314000, stop = 103435000) -> myMgl}

}
\seealso{
Other Build list: \code{\link{buildFromEnsgs}},
  \code{\link{buildFromNames}}
}

