% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exMgl.R
\name{exMgl}
\alias{exMgl}
\title{exMgl}
\usage{
exMgl(saveDownload = F)
}
\arguments{
\item{saveDownload}{A logical vector indicating if the data should be saved as 'myMgl.rda'}
}
\description{
\code{exMgl} returns example mgl list for vignette myMgl
}
\details{
Example of fully built list for five candidate genes(note: element 20 - AEI is still missing as this data requires IRB approval and access via dbGaP).
}
\examples{
exMgl() -> myMgl

}

