% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSnpIds_function.R
\name{fixSnpIds}
\alias{fixSnpIds}
\title{Convert chr_pos to rs id}
\usage{
fixSnpIds(mgl)
}
\arguments{
\item{mgl}{List; see \code{\link{buildFromNames}}, \code{\link{buildFromRegion}}, or \code{\link{buildFromEnsgs}}}
}
\description{
\code{fixSnpIds} converts chr_pos identifiers to rs ids for cisEqtl data released from GTEx
}
\examples{
\dontrun{fixSnpIds(myMgl) -> myMgl}

}
\seealso{
\code{\link{addCisEqtl}}
}

