% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSeparate_function.R
\name{listSeparate}
\alias{listSeparate}
\title{Returns an individual element for each gene}
\usage{
listSeparate(mgl, element = c(""))
}
\arguments{
\item{mgl}{List; see \code{\link{buildFromNames}}, \code{\link{buildFromRegion}}, or \code{\link{buildFromEnsgs}}}

\item{element}{A numeric indicating which element to be isolated}
}
\description{
\code{listSeparate} returns a list of any given element
}
\examples{
listSeparate(myMgl, element = 3) -> sep

}
\seealso{
Other list: \code{\link{listElements}}
}

