% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDnaseSig_function.R
\name{makeDnaseSig}
\alias{makeDnaseSig}
\title{Filter Dnase (Maurano) Results}
\usage{
makeDnaseSig(mgl)
}
\arguments{
\item{mgl}{List; see \code{\link{buildFromNames}}, \code{\link{buildFromRegion}}, or \code{\link{buildFromEnsgs}}}
}
\description{
\code{makeDnaseSig} returns a list with length corresponding to the number of candidate genes.  Each element is a dataframe of all the Dnase results filtered for those that are significant.
}
\examples{
myMgl <- makeDnaseSig(myMgl)

}
\seealso{
Other output: \code{\link{makeAeiPlot}},
  \code{\link{makeCoXpGene}},
  \code{\link{makeCoXpTranscript}},
  \code{\link{makeGoSearch}}, \code{\link{makeGo}},
  \code{\link{makeMultiEqtl}},
  \code{\link{makeOverlapTable}},
  \code{\link{makeOverlap}},
  \code{\link{makePhenotypeSearch}},
  \code{\link{makePhenotypes}},
  \code{\link{makeSnpSearch}}, \code{\link{makeSnps}},
  \code{\link{makeSummary}}
}

