% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMgl.R
\docType{data}
\name{myMgl}
\alias{myMgl}
\title{myMgl}
\format{An 'mgl' with 20 elements}
\usage{
myMgl
}
\description{
Example of fully built list (note: element 20 - AEI is still missing as this data requires IRB approval and access via dbGaP)
}
\keyword{datasets}

